client request implemented

+++ networking.hpp
@@ -1,13 +1,14 @@
 #ifndef __NETWORKING_HPP__
 #define __NETWORKING_HPP__
 
+#include <cstdint>
 #include "protocol.hpp"
 
 namespace net {
 
     struct tcp_server final {
         const char* address;
-        int port;
+        uint16_t port;
 
         void listen() const;
     };
